requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=0.19.0
langchain-community>=0.3.26
langchain[openai]>=0.3.26
langgraph>=0.4.9
langsmith>=0.4.1
openai>=1.91.0
pymysql>=1.1.1
sqlalchemy>=2.0.41
unstructured>=0.18.3
openpyxl>=3.1.5
networkx>=3.5
langchain-openai>=0.3.25
langchain-pinecone>=0.2.8
faiss-cpu>=1.11.0
pydantic>=2.0.0
typing-extensions>=4.8.0 